requirements.txt
Pillow
shapely
matplotlib
numpy
tensorflow
opencv-python